third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-stack-overflow-source-url.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script src="../inspector/console-test.js"></script>
<script type="text/javascript">
function onload()
{
    if (window.testRunner)
        testRunner.waitUntilDone();
    setTimeout(showInspectorAndRunTest, 0);

    var scriptContents = "function boo(){ \n boo(); \n } \n function foo() { \n boo();} \n //# sourceURL=foo2.js";
    eval(scriptContents);
    foo();
}


function showInspectorAndRunTest()
{
    if (window.testRunner)
        testRunner.showWebInspector();
    runTest();
}

function test() {
  ConsoleTestRunner.expandConsoleMessages();
  ConsoleTestRunner.dumpConsoleMessages();
  TestRunner.addResult('TEST COMPLETE');
  TestRunner.completeTest();
}
</script>
</head>
<body onload="onload()">
<p>Tests that when stack overflow happens before inspector is open source url is correctly shown in console.</p>
</body>
</html>
