third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-stack-overflow-source-url-expected.txt

CONSOLE ERROR: line 1: Uncaught RangeError: Maximum call stack size exceeded
Tests that when stack overflow happens before inspector is open source url is correctly shown in console.

foo2.js:1 Uncaught RangeError: Maximum call stack size exceeded
    at boo (foo2.js:1:13)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
    at boo (foo2.js:2:2)
TEST COMPLETE
